[refactor] Convert trial_stim into subfeature
